# perf: do not `compose` if it has only one handler