Made it does not `compose` if it has only one handler. `await composed` is slow and could be speed up by avoiding `compose`.

## Benchmark

Will be faster if it returns only "Hello World".

### handle-event

17%  Faster!!

```
hono - trie-router(default) x 421,632 ops/sec ±4.36% (79 runs sampled)
hono - regexp-router x 452,187 ops/sec ±5.90% (69 runs sampled)
hono - trie-router(default) - New x 506,280 ops/sec ±3.92% (85 runs sampled)
hono - regexp-router - New x 610,975 ops/sec ±4.52% (83 runs sampled)
```

### Deno - flash

Not much has changed.

```
v2.1.3
Requests/sec: 185881.3210
New
Requests/sec: 188679.7531
```

### Bun

11% Faster!!

```
v2.1.3
Requests/sec: 133396.6236
New
Requests/sec: 149045.7514
```