@usualoma 2022-09-01
@yusukebe 

Awesome! But I think the logger is called twice in the following cases.

```ts
app.use('*', logger())
app.get('/', (c) => c.text('Hono!!'))
```

```
GET /not-found
```