@yusukebe 2022-09-01
Hi @usualoma !

Thanks for pointing that out. It is indeed a bug. I have fixed it on this PR, can you review it?

#494